pkg/recommender/requirements.txt
black==23.1.0
click==8.1.3
grpcio==1.51.1
grpcio-tools==1.51.1
joblib==1.2.0
mypy-extensions==0.4.3
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==2.6.2
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.1
scipi==3.0
scipy==1.10.0
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
typing-extensions==4.4.0
